Read this It will HELP A LOT
This is the code for finding whether the number given from user is prime or not.
For rookies(who are new to JAVA Language) this will be really hard for them to understand on early stages of coding. It was same for me but, I did more practice on this Prime code to create better understanding. #2-3 times writing the same code while describing meaning of every line to yourself will really help tu create a good logic for JAVA or any programming language.
Just talk to yourself like a teacher teaching you coding while writing the code. That's it and you are done..!!